requirements.txt
torch
torchvision
torchaudio
numpy
pillow
opencv-python
ffmpeg-python
redis
onnxruntime
insightface
gfpgan
